Extract manga badge components
